Remove redundant casts all over more generics

+++ src/syndie/db/LoginManager.java
@@ -116,7 +116,7 @@ public class LoginManager extends CommandImpl {
                 // Store the default author into nym preferences early on,
                 // preventing an NPE later (specifically: when we post our first message
                 // with default settings to our own forum).
-                long authorId = ((Long)client.getNymChannels().getIdentityChannelIds().get(0)).longValue();
+                long authorId = client.getNymChannels().getIdentityChannelIds().get(0).longValue();
                 Hash author = client.getChannelHash(authorId);
                 Properties prefs = client.getNymPrefs();
                 prefs.setProperty("editor.defaultAuthor", author.toBase64());
